+++ src/mock/data.js
@@ -1,33 +1,10 @@
+import { getRandomInteger } from "../utils/utils.js";
+import { shuffle } from "../utils/utils.js";
+
 const DESCRIPTION_LENGTH = 5;
 const MIN_PRICE = 50;
 const MAX_PRICE = 200;
-/**
- * Возвращает случайное число из диапазона
- * @param  {number} a=0
- * @param  {number} b=1
- * @return число
- */
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
 
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-/**
- * Меняет порядок элементов в массиве
- * @param  {array} arr
- * @return массив
- */
-const shuffle = (arr) => {
-  var j, x, i;
-  for (i = arr.length - 1; i > 0; i--) {
-    j = Math.floor(Math.random() * (i + 1));
-    x = arr[i];
-    arr[i] = arr[j];
-    arr[j] = x;
-  }
-  return arr;
-};
 /**
  */
 const generateTripType = () => {
@@ -57,16 +34,16 @@ const generateOfferType = () => {
 
 const generateDescription = () => {
   const sentences = [
-    `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
-    `Cras aliquet varius magna, non porta ligula feugiat eget.`,
-    `Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra.`,
-    `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.`,
-    `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
-    `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`,
-    `Sed sed nisi sed augue convallis suscipit in sed felis.`,
-    `Aliquam erat volutpat.`,
-    `Nunc fermentum tortor ac porta dapibus.`,
-    `In rutrum ac purus sit amet tempus.`
+    `Lorem ipsum dolor sit amet, consectetur adipiscing elit. `,
+    `Cras aliquet varius magna, non porta ligula feugiat eget. `,
+    `Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. `,
+    `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. `,
+    `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. `,
+    `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. `,
+    `Sed sed nisi sed augue convallis suscipit in sed felis. `,
+    `Aliquam erat volutpat. `,
+    `Nunc fermentum tortor ac porta dapibus. `,
+    `In rutrum ac purus sit amet tempus. `
   ];
   return shuffle(sentences).slice(0, DESCRIPTION_LENGTH);
 };
@@ -87,6 +64,18 @@ const generatePrice = () => {
   return getRandomInteger(MIN_PRICE,MAX_PRICE);
 };
 
+const generateStartTime = () => {
+  const timeValue = [`12:00`, `12:30`, `13:00`, `13:30`, `14:00`, `14:30`];
+  const randomIndex = getRandomInteger(0, timeValue.length - 1);
+  return timeValue[randomIndex];
+};
+
+const generateEndTime = () => {
+  const timeValue = [`15:30`, `16:00`, `16:30`, `17:00`, `17:30`, `18:00`];
+  const randomIndex = getRandomInteger(0, timeValue.length - 1);
+  return timeValue[randomIndex];
+};
+
 const generateTripData = () => {
   return {
     type: generateTripType(),
@@ -95,6 +84,8 @@ const generateTripData = () => {
     description: generateDescription(),
     photo: generatePictures(),
     price : generatePrice(),
+    startTime: generateStartTime(),
+    endTime: generateEndTime(),
   };
 };
