+++ src/view/formOffer.js
@@ -1,0 +1,75 @@
+const createFormOfferTemplate = (pointDetails) => {
+  const {description, pictures} = pointDetails;
+
+  return (
+    `<section class="event__details">
+    <section class="event__section  event__section--offers">
+      <h3 class="event__section-title  event__section-title--offers">Offers</h3>
+
+      <div class="event__available-offers">
+        <div class="event__offer-selector">
+          <input class="event__offer-checkbox  visually-hidden" id="event-offer-luggage-1" type="checkbox" name="event-offer-luggage" checked="">
+          <label class="event__offer-label" for="event-offer-luggage-1">
+            <span class="event__offer-title">Add luggage</span>
+            +
+            €&nbsp;<span class="event__offer-price">30</span>
+          </label>
+        </div>
+
+        <div class="event__offer-selector">
+          <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-1" type="checkbox" name="event-offer-comfort" checked="">
+          <label class="event__offer-label" for="event-offer-comfort-1">
+            <span class="event__offer-title">Switch to comfort class</span>
+            +
+            €&nbsp;<span class="event__offer-price">100</span>
+          </label>
+        </div>
+
+        <div class="event__offer-selector">
+          <input class="event__offer-checkbox  visually-hidden" id="event-offer-meal-1" type="checkbox" name="event-offer-meal">
+          <label class="event__offer-label" for="event-offer-meal-1">
+            <span class="event__offer-title">Add meal</span>
+            +
+            €&nbsp;<span class="event__offer-price">15</span>
+          </label>
+        </div>
+
+        <div class="event__offer-selector">
+          <input class="event__offer-checkbox  visually-hidden" id="event-offer-seats-1" type="checkbox" name="event-offer-seats">
+          <label class="event__offer-label" for="event-offer-seats-1">
+            <span class="event__offer-title">Choose seats</span>
+            +
+            €&nbsp;<span class="event__offer-price">5</span>
+          </label>
+        </div>
+
+        <div class="event__offer-selector">
+          <input class="event__offer-checkbox  visually-hidden" id="event-offer-train-1" type="checkbox" name="event-offer-train">
+          <label class="event__offer-label" for="event-offer-train-1">
+            <span class="event__offer-title">Travel by train</span>
+            +
+            €&nbsp;<span class="event__offer-price">40</span>
+          </label>
+        </div>
+      </div>
+    </section>
+
+    <section class="event__section  event__section--destination">
+      <h3 class="event__section-title  event__section-title--destination">Destination</h3>
+      <p class="event__destination-description">${description.join(``)}</p>
+
+      <div class="event__photos-container">
+        <div class="event__photos-tape">
+          <img class="event__photo" src="img/photos/1.jpg" alt="Event photo">
+          <img class="event__photo" src="img/photos/2.jpg" alt="Event photo">
+          <img class="event__photo" src="img/photos/3.jpg" alt="Event photo">
+          <img class="event__photo" src="img/photos/4.jpg" alt="Event photo">
+          <img class="event__photo" src="img/photos/5.jpg" alt="Event photo">
+        </div>
+      </div>
+    </section>
+  </section>`
+  );
+};
+
+export {createFormOfferTemplate};
